fixup! JAMES-2725 Split object storage tests per implementation
