+++ server/container/guice/cassandra-rabbitmq-guice/src/test/java/org/apache/james/WithoutSwiftOrAwsS3Test.java
@@ -19,23 +19,8 @@
 
 package org.apache.james;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.apache.james.blob.objectstorage.AESPayloadCodec;
-import org.apache.james.blob.objectstorage.PayloadCodec;
-import org.apache.james.modules.objectstorage.aws.s3.DockerAwsS3TestRule;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-@ExtendWith(WithEncryptedAwsS3Extension.class)
-public class WithEncryptedAwsS3 implements JmapJamesServerContract, MailsShouldBeWellReceived, JamesServerContract {
-
-    @Test
-    void encryptedPayloadShouldBeConfiguredWhenProvidingEncryptedPayloadConfiguration(GuiceJamesServer jamesServer) {
-        PayloadCodec payloadCodec = jamesServer.getProbe(DockerAwsS3TestRule.TestAwsS3BlobStoreProbe.class)
-            .getAwsS3PayloadCodec();
-
-        assertThat(payloadCodec)
-            .isInstanceOf(AESPayloadCodec.class);
-    }
+@ExtendWith(WithoutSwiftOrAwsS3Extension.class)
+public class WithoutSwiftOrAwsS3Test implements JmapJamesServerContract, MailsShouldBeWellReceived, JamesServerContract {
 }
